FrequencySpinBox: give up-to-date value
With keyboardTracking disabled, currValue and spin box's value were
only getting updated once the spin box lost focus. This was problematic
for times when a button was pressed to perform an action while the
FrequencySpinBox was still in focus and a new value had been entered,
as the old value was returned. This caused wonky behaviour when changing
frequencies in the Panoramic Spectrum dialog of SigDigger, and then
starting the scan without first clicking away from a FrequencySpinBox.

The new logic always gives an up-to-date value when value() is called,
while still avoiding prematurely issuing valueChanged() signals.