GLWaterfall: better handle short averages
Resetting the signal to 0 dB and applying what's essentially an IIR low
pass filter (not averaging) was a terrible strategy, as it would produce
very incorrect overly strong signals for short averaging durations.

At least in the new logic, it starts off with the first block of data,
and then applies the IIR filter relative to that. Perhaps real averaging
would be better than the IIR filter, but that's another day's task.