AbstractWaterfall: plot with devicePixelRatio of 1
This is around 20x faster while also showing twice as many points, and
would be around 40x faster with the same number of points. A one pixel
wide line is still easy to see on a high DPI display.

This change also always draws the spectrum plot as HDPI on capable
displays, regardless of the waterfall's HDPI awareness.